% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-to-text.r
\name{html_to_text}
\alias{html_to_text}
\title{Convert HTML to Text}
\usage{
html_to_text(x, include_attributes = FALSE)
}
\arguments{
\item{x}{HTML content (length 1 character vector)}

\item{include_attributes}{Sets whether any attribute values are included in the output}
}
\value{
character vector of cleaned text
}
\description{
Convert HTML to Text
}
\examples{
fils <- list.files(system.file("extdata", package="jericho"), full.names=TRUE)
for(f in fils) {
  d <- readBin(f, "raw", file.size(f))
  d <- rawToChar(d)
  html_to_text(d)
}
}
