% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jericho-package.R
\docType{package}
\name{jericho}
\alias{jericho}
\alias{jericho-package}
\title{Break Down the Walls of 'HTML' Tags into Usable Text}
\description{
Structured 'HTML' content can be useful when you need to parse data tables or
other tagged data from within a document. However, it is also useful to obtain
"just the text" from a document free from the walls of tags that surround it.
Tools are provied that wrap methods in the 'Jericho HTML Parser' Java library
by Martin Jericho \url{http://jericho.htmlparser.net/docs/index.html}. Martin's
library is used in many at-scale projects, icluding the 'The Internet Archive'.
}
\author{
Bob Rudis (bob@rud.is)
}
