% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-to-text.r
\name{render_html_to_text}
\alias{render_html_to_text}
\title{Render HTML to Text}
\usage{
render_html_to_text(x)
}
\arguments{
\item{x}{HTML content (length 1 character vector)}
}
\value{
character vector of cleaned text
}
\description{
This provides a human readable version of the segment content that is modelled on
the way Mozilla Thunderbird and other email clients provide an automatic conversion
of HTML content to text in their alternative MIME encoding of emails.
}
